# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Coworking API",
    "description": "Автоматически сгенерированная документация",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/coworking-centers/": {
      "get": {
        "summary": "Получить все коворкинг-центры",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Создать новый коворкинг-центр",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "example": "any"
                },
                "latitude": {
                  "example": "any"
                },
                "longitude": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "openingTime": {
                  "example": "any"
                },
                "closingTime": {
                  "example": "any"
                },
                "amenities": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/coworking-centers/all": {
      "get": {
        "summary": "Получить все центры (включая неактивные) для менеджеров",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/coworking-centers/{id}": {
      "get": {
        "summary": "Получить конкретный коворкинг-центр",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Обновить коворкинг-центр",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "example": "any"
                },
                "latitude": {
                  "example": "any"
                },
                "longitude": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "openingTime": {
                  "example": "any"
                },
                "closingTime": {
                  "example": "any"
                },
                "amenities": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Удалить коворкинг-центр",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/bookings/coworking/{coworkingCenterId}": {
      "get": {
        "summary": "Получить бронирования для коворкинг-центра",
        "description": "",
        "parameters": [
          {
            "name": "coworkingCenterId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/discounts/": {
      "get": {
        "summary": "Получить все скидки",
        "description": "",
        "parameters": [
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Создать скидку",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "percentage": {
                  "example": "any"
                },
                "maxDiscountAmount": {
                  "example": "any"
                },
                "usageLimit": {
                  "example": "any"
                },
                "startDate": {
                  "example": "any"
                },
                "endDate": {
                  "example": "any"
                },
                "applicableDays": {
                  "example": "any"
                },
                "applicableHours": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/discounts/{id}": {
      "get": {
        "summary": "Получить конкретную скидку",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Обновить скидку",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "percentage": {
                  "example": "any"
                },
                "maxDiscountAmount": {
                  "example": "any"
                },
                "usageLimit": {
                  "example": "any"
                },
                "startDate": {
                  "example": "any"
                },
                "endDate": {
                  "example": "any"
                },
                "applicableDays": {
                  "example": "any"
                },
                "applicableHours": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Удалить скидку",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/discounts/status/active": {
      "get": {
        "summary": "Получить активные скидки",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/discounts/check/availability": {
      "get": {
        "summary": "Проверить доступность скидки по критериям",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "time",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dayOfWeek",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}